[test_nn] move dropout tests to test/nn/test_dropout.py
